fix: Rename file and const names to maintain a standard
